Added tagindexviewmodel to include useagecnt and list of articles and mediakit files. Updated tagcontroller index and edit to use tagindexviewmodel. updated index,delete and edit view for tag.

+++ src/NewsSite/Controllers/OwnersController.cs
@@ -240,7 +240,10 @@ namespace NewsSite.Controllers
                 _context.ArticleMediaKitFile.RemoveRange(_context.ArticleMediaKitFile.Where(m => m.MediaKitFileId == mediaKitFile.MediaKitFileId));
                 _context.MediaKitFile.Remove(mediaKitFile);
                 string filename = hostingEnv.WebRootPath + $@"\mediakitfiles\{mediaKitFile.URL}";
-                System.IO.File.Delete(filename);
+                if (System.IO.File.Exists(filename))
+                {
+                    System.IO.File.Delete(filename);
+                }
             }
 
             _context.Owner.Remove(owner);
